fix catch wrong eaten dish cause wrong query in EatenDishDao

+++ app/src/main/java/com/example/health/screens/main/diary/compose/DetailFoodLayout.kt
@@ -14,9 +14,6 @@ import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.Button
-import androidx.compose.material.ButtonDefaults
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 
@@ -40,7 +37,7 @@ fun DetailFoodLayout(
     carb: Float,
     fat: Float,
     protein: Float,
-    quantity: Int,             // ⬅️ Thêm dòng này
+    quantity: Float,             // ⬅️ Thêm dòng này
     quantityType: String
 ) {
     Box(modifier = Modifier.fillMaxSize()) {
